mcp_servers/web_search/requirements.txt
fastmcp>=3.2.4
httpx>=0.28.1
pyjwt>=2.12.1
